+++ config/graphql/graphene_types.py
@@ -82,6 +82,8 @@ from config.graphql.pipeline_types import (  # noqa: F401
     PipelineComponentsType,
     PipelineComponentType,
     PipelineSettingsType,
+    StageCoverageType,
+    SupportedMimeTypeType,
 )
 from config.graphql.social_types import (  # noqa: F401
     BadgeDistributionType,